Enhanced CustomCard layout and design
The CustomCard widget in our Flutter project has been updated to provide
a more informative and visually appealing interface. The spacing of
elements within the card has been adjusted by increasing the SizedBox
height, which enhances content separation and readability. A new feature
displays a series of profile images to indicate social proof or user
engagement. Additionally, a '4 Star' TextButton has been introduced to
highlight the quality rating of the card's subject, along with a text
element showing the number of reviews. The inclusion of ratings and
reviews adds a layer of user interaction and trustworthiness to the
card's content.