Refactor home screen and enable logout
The home screen was simplified by removing the location-fetching
FutureBuilder and replaced it with a bottom navigation-controlled page
view. This aligns with the modular design pattern, facilitating easier
navigation and maintenance.

Additionally, the logout functionality on the profile screen was
implemented. It now properly utilizes the AuthService to sign the user
out and redirects to the SignIn screen, enhancing the app's security and
user experience by ensuring clean session termination.