Update app icon and remove outdated assets
Replaced the app icon with a new `launcher_icon` to refresh the app's
branding and deleted a series of unused avatar, background, and marker
image assets to declutter the project. Additionally, performed a
downgrade of several dependencies as outlined in the `pubspec.lock` to
align with compatibility requirements or resolve conflicts after testing
the newer versions. Updated `pubspec.yaml` to reflect these changes in
launcher icons configuration.