Enhance profile UI with friend features
Introduced a floating action button on the profile screen, enabling
users to navigate to a new Friend Requests page. Additionally, added
information on friends count to the profile view to enhance user
engagement. The Background widget was encapsulated within a Scaffold for
better structure and to support the new floating button. Also updated
the routing information to acknowledge the Friend Requests screen. These
improvements aim to increase user interactivity and provide a more
social experience within the app.